added fastify autoload and moved db plugin to util directory
