+++ package.json
@@ -28,6 +28,7 @@
   },
   "dependencies": {
     "@casl/ability": "^6.7.3",
+    "@fastify/autoload": "^6.3.1",
     "@fastify/jwt": "^10.0.0",
     "ajv": "^8.17.1",
     "axios": "^1.11.0",
